vCreated the thread route with endpoints for creating, retrieving and deleting threads. Added functionality to generate unique usernames for users on registration. Also added middleware to verify tokens and did some general code refactoring.
