+++ routes/PostRoutes.js
@@ -1,22 +1,23 @@
 // importing necessary libraries
 const express = require('express')
 const Router = express.Router()
-const {makePosts,getPostById,updatePost,deletePost} = require('../controllers/posts')
-
+const {
+  makePosts,
+  getPostById,
+  updatePost,
+  deletePost,
+} = require('../controllers/postController')
 
 //creating endpoint for posts
 Router.route('/').post(makePosts)
-
 
 // //endpoint for getting all posts
 Router.route('/:id').get(getPostById)
 
 //updating endpoint for posts
-Router.route('/:d').post(updatePost)
+Router.route('/:id').put(updatePost)
 
 //deleting endpoint for posts
-Router.route('/:id').post(deletePost)
-
-
+Router.route('/:id').delete(deletePost)
 
 module.exports = Router
